pyproject.toml
[tool.poetry]
name = "simple-gcp-rag"
version = "0.1.0"
description = ""
authors = ["Kamil Zieliński <kamilkk.zielinski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
google-cloud-aiplatform = "^1.71.0"
vertexai = "^1.71.1"
tqdm = "^4.67.1"
proto-plus = "1.24.0.dev1"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
